requirements.txt
Flask==2.3.2
requests==2.26.0
openai==1.6.1
python-dotenv==1.0.0
flask_cors==4.0.0
virtualenv==20.25.0
aiohttp==3.9.1
torch
pybase64
numpy
surya-ocr